fix a bug where sdk responses can only be json_encoded once
